Feat: moved state to app file, removed repeat interfaces, added user name to auth state
