+++ src/app/state/app.state.ts
@@ -1,0 +1,10 @@
+// import { AuthState } from './auth/auth.reducer';
+
+// export interface State {
+//   user: AuthState;
+// }
+// // The default state when the app starts -- no token, no name
+// export const initialState: AuthState = {
+//   token: null,
+//   name: null
+// };
