+++ src/app/state/auth/auth.actions.ts
@@ -1,0 +1,10 @@
+import { createAction, props } from '@ngrx/store';
+
+export const login = createAction(
+  '[Auth] Login',
+  props<{ token: string, name: string }>()
+);
+
+export const logout = createAction(
+  '[Auth] Logout'
+);
