+++ src/app/state/auth/auth.selectors.ts
@@ -2,21 +2,17 @@ import { createFeatureSelector, createSelector } from '@ngrx/store';
 import { AuthState } from './auth.reducer';
 
 export const tokenKey = 'auth';
-// Select the entire "auth" state
 export const selectAuthState = createFeatureSelector<AuthState>(tokenKey);
 
-// Used to query the store and retrieve the auth token
 export const selectToken = createSelector(
   selectAuthState,
-  (state) => state.token
+  (state: AuthState) => state.token
 );
 
-// Check if the user is authenticated
-// Returns true if a token exists, otherwise false
-export const selectIsAuthenticated = createSelector(
+export const selectName = createSelector(
   selectAuthState,
-  (state) => !!state.token
-);
+  (state: AuthState) => state.name
+)
 
 // Used in the component like this:
 // this.store.select(selectToken).subscribe(token => console.log(token));
